requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cbor2==5.7.1
certifi==2025.10.5
click==8.3.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
frozenlist==1.8.0
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modal==1.2.1
multidict==6.7.0
openai==2.7.1
propcache==0.4.1
protobuf==6.33.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
sentry-sdk==2.43.0
shellingham==1.5.4
sigtools==4.0.1
sniffio==1.3.1
starlette==0.49.3
synchronicity==0.10.2
toml==0.10.2
tqdm==4.67.1
typer==0.20.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
